+++ c/approaching.c
@@ -1,0 +1,15 @@
+#include <stdio.h>
+
+int main(void)
+{
+	int x = 9;
+
+	while (x \
+		  \
+		   \
+		    --> 0) {
+		printf("x = %d\n", x);
+	}
+
+	return 0;
+}