+++ hpc/nvidia/cuda/samples/cudaGetDeviceProperties.cu
@@ -2,18 +2,27 @@
  * cuda-12.9 struct cudaDeviceProp: commit d072d32dcb00 ("cuda: cudaGetDeviceProperties.cu: display more")
  */
 #include <stdio.h>
+#if defined(HAVE_HCCL)
+#include <hc_runtime.h>
+#include "hpcc_helpers.h"
+#else
 #include <cuda_runtime.h>
 #include "cuda_helpers.h"
+#endif
 
 
 int main(void)
 {
 	int i;
+#ifdef HAVE_HCCL
+	hcDeviceProp_t prop;
+	hc_init(0);
+	hcGetDeviceProperties(&prop, 0);
+#else
 	cudaDeviceProp prop;
-
 	cuda_init(0);
-
 	cudaGetDeviceProperties(&prop, 0);
+#endif
 
 	printf("name %s\n", prop.name);