+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+- metax: add hcGetDeviceProperties.hpcc, symlink dim3.hpcc;
+- cuda: add dim3.cu;
 #### Changed
 #### Deprecated
 #### Removed
@@ -28,9 +30,9 @@ and this project adheres to
 - make: add filter.mk;
 - wayland: add info.sh;
 - X11/Xorg: Introduce;
-- metax: add ht-envs.sh to set ENVs of hpcc, add abbreviation.md, parallel.hpcc, hcclGetVersion.hpcc, `hpcc_helpers.hpcc`, hcMalloc.hpcc, matrix-mul.hpcc, hcGetDeviceProperties.hpcc, symlink dim3.hpcc;
+- metax: add ht-envs.sh to set ENVs of hpcc, add abbreviation.md, parallel.hpcc, hcclGetVersion.hpcc, `hpcc_helpers.hpcc`, hcMalloc.hpcc, matrix-mul.hpcc;
 - hpc: add metax/ directory;
-- cuda: add driver-version.c from ostools, add parallel.cu, cudaMalloc.cu, dim3.cu, `cuda_helpers.cu`, metax symlink matrix-mul.cu;
+- cuda: add driver-version.c from ostools, add parallel.cu, cudaMalloc.cu, `cuda_helpers.cu`, metax symlink matrix-mul.cu;
 - compiler: Introduce OpenACC directory;
 - ai: Introduce AI directory;
 - git: add push-remote-all.sh, rm-permanent.sh from ostools;