+++ CHANGELOG.md
@@ -13,9 +13,9 @@ and this project adheres to
 - mk: All targets use small case words, this is a big change;
 #### Added
 - ascend: add abbreviation.md;
-- metax: add hcGetDeviceProperties.hpcc, cuda2hccl.h, hcDeviceGetAttribute.hpcc, hcDeviceGetP2PAttribute.hpcc, blas.hpcc, symlink dim3.hpcc, matrix-mul-int8.hpcc, matrix-mul-fp16.hpcc;
+- metax: add hcGetDeviceProperties.hpcc, cuda2hccl.h, hcDeviceGetAttribute.hpcc, hcDeviceGetP2PAttribute.hpcc, blas.hpcc, symlink dim3.hpcc, matrix-mul-int8.hpcc, matrix-mul-fp16.hpcc, rand.hpcc;
 - cuda: add dim3.cu, symlink cudaDeviceGetP2PAttribute.cu, matrix-mul.sh, matrix-mul-double.cu, matrix-mul-int8.cu, blas.cu;
-- nvidia: add README.Nsight.md from ostools;
+- nvidia: add README.Nsight.md from ostools, add rand.cu;
 #### Changed
 #### Deprecated
 #### Removed